Move show_docs to its own module
