+++ mkosi/__init__.py
@@ -55,7 +55,6 @@ from mkosi.config import (
     Compression,
     Config,
     ConfigFeature,
-    DocFormat,
     JsonEncoder,
     KeySourceType,
     ManifestFormat,
@@ -76,6 +75,7 @@ from mkosi.config import (
 )
 from mkosi.context import Context
 from mkosi.distributions import Distribution
+from mkosi.documentation import show_docs
 from mkosi.installer import clean_package_manager_metadata
 from mkosi.kmod import gen_required_kernel_modules, loaded_modules, process_kernel_modules
 from mkosi.log import ARG_DEBUG, complete_step, die, log_notice, log_step
@@ -90,7 +90,6 @@ from mkosi.run import (
     chroot_options,
     finalize_interpreter,
     finalize_passwd_mounts,
-    find_binary,
     fork_and_wait,
     run,
 )
@@ -3584,47 +3583,6 @@ def bump_image_version() -> None:
 
     logging.info(f"Bumping version: '{version}' → '{new_version}'")
     version_file.write_text(f"{new_version}\n")
-
-
-def show_docs(args: Args, *, resources: Path) -> None:
-    if args.doc_format == DocFormat.auto:
-        formats = [DocFormat.man, DocFormat.pandoc, DocFormat.markdown, DocFormat.system]
-    else:
-        formats = [args.doc_format]
-
-    manual = args.cmdline[0] if args.cmdline else "mkosi"
-
-    while formats:
-        form = formats.pop(0)
-        try:
-            if form == DocFormat.man:
-                man = resources / f"man/{manual}.1"
-                if not man.exists():
-                    raise FileNotFoundError()
-                run(["man", "--local-file", man])
-                return
-            elif form == DocFormat.pandoc:
-                if not find_binary("pandoc"):
-                    logging.error("pandoc is not available")
-                pandoc = run(
-                    ["pandoc", "-t", "man", "-s", resources / f"man/{manual}.md"],
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.DEVNULL,
-                    log=False,
-                )
-                run(["man", "--local-file", "-"], input=pandoc.stdout)
-                return
-            elif form == DocFormat.markdown:
-                page((resources / f"man/{manual}.md").read_text(), args.pager)
-                return
-            elif form == DocFormat.system:
-                run(["man", manual], log=False)
-                return
-        except (FileNotFoundError, subprocess.CalledProcessError) as e:
-            if not formats:
-                if isinstance(e, FileNotFoundError):
-                    die("The mkosi package does not contain the man page {manual:r}.")
-                raise e
 
 
 def expand_specifier(s: str) -> str:
